File download

+++ client/src/main/java/it/unishare/client/managers/DatabaseManager.java
@@ -1,13 +1,12 @@
-package it.unishare.client.database;
+package it.unishare.client.managers;
 
-import it.unishare.client.connection.ConnectionManager;
 import it.unishare.client.utils.Settings;
 import it.unishare.common.connection.kademlia.KademliaFile;
 import it.unishare.common.connection.kademlia.KademliaFileData;
+import it.unishare.common.models.User;
 
 import java.io.File;
 import java.nio.file.Files;
-import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.sql.*;
 import java.util.ArrayList;
@@ -41,7 +40,7 @@ public class DatabaseManager {
 
 
     /**
-     * Get connection to the SQLite database
+     * Get managers to the SQLite database
      *
      * @return  database {@link Connection}
      */
@@ -80,6 +79,7 @@ public class DatabaseManager {
     private void initializeDatabase() {
         String sql = "CREATE TABLE IF NOT EXISTS my_files(" +
                     "id INTEGER PRIMARY KEY AUTOINCREMENT," +
+                    "user_id INTEGER NOT NULL," +
                     "key BLOB NOT NULL," +
                     "title TEXT NOT NULL," +
                     "university TEXT NOT NULL," +
@@ -103,17 +103,21 @@ public class DatabaseManager {
     /**
      * Get all files
      */
-    public List<KademliaFile> getAllFiles() {
+    public List<KademliaFile> getUserFiles(User user) {
         List<KademliaFile> result = new ArrayList<>();
-        String sql = "SELECT key, title, university, department, course, teacher FROM my_files";
+        String sql = "SELECT key, title, university, department, course, teacher FROM my_files WHERE user_id = ?";
 
         try {
             Connection connection = getConnection();
-            Statement statement = connection.createStatement();
-            ResultSet resultSet = statement.executeQuery(sql);
+
+            PreparedStatement statement = connection.prepareStatement(sql);
+            statement.setLong(1, user.getId());
+
+            ResultSet resultSet = statement.executeQuery();
 
             while (resultSet.next()) {
                 KademliaFileData data = new KademliaFileData(
+                        user.getFullName(),
                         resultSet.getString("title"),
                         resultSet.getString("university"),
                         resultSet.getString("department"),
@@ -130,6 +134,8 @@ public class DatabaseManager {
                 result.add(file);
             }
 
+            connection.close();
+
         } catch (SQLException e) {
             e.printStackTrace();
         }
@@ -143,19 +149,20 @@ public class DatabaseManager {
      *
      * @param   file    file
      */
-    public void addFile(KademliaFile file) {
-        String sql = "INSERT INTO my_files(key, title, university, department, course, teacher) VALUES(?,?,?,?,?,?)";
+    public void addFile(User user, KademliaFile file) {
+        String sql = "INSERT INTO my_files(user_id, key, title, university, department, course, teacher) VALUES(?, ?, ?, ?, ?, ?, ?)";
 
         try {
             Connection connection = getConnection();
 
             PreparedStatement statement = connection.prepareStatement(sql);
-            statement.setBytes(1, file.getKey().getBytes());
-            statement.setString(2, file.getData().getTitle());
-            statement.setString(3, file.getData().getUniversity());
-            statement.setString(4, file.getData().getDepartment());
-            statement.setString(5, file.getData().getCourse());
-            statement.setString(6, file.getData().getTeacher());
+            statement.setLong(1, user.getId());
+            statement.setBytes(2, file.getKey().getBytes());
+            statement.setString(3, file.getData().getTitle());
+            statement.setString(4, file.getData().getUniversity());
+            statement.setString(5, file.getData().getDepartment());
+            statement.setString(6, file.getData().getCourse());
+            statement.setString(7, file.getData().getTeacher());
 
             statement.executeUpdate();
             connection.close();
@@ -170,14 +177,15 @@ public class DatabaseManager {
      *
      * @param   key     key
      */
-    public void deleteFile(byte[] key) {
-        String sql = "DELETE FROM my_files WHERE key = ?";
+    public void deleteFile(long userId, byte[] key) {
+        String sql = "DELETE FROM my_files WHERE user_id = ? AND key = ?";
 
         try {
             Connection connection = getConnection();
 
             PreparedStatement statement = connection.prepareStatement(sql);
-            statement.setBytes(1, key);
+            statement.setLong(1, userId);
+            statement.setBytes(2, key);
 
             statement.executeUpdate();
             connection.close();
